Worked more on networking. Started filing in some areas for AI. Worked on the gameloop.

+++ DungeonMaster/src/Enemies/Enemy.java
@@ -4,6 +4,6 @@ import dungeonmaster.NonPlayerCharacter;
 
 public class Enemy extends NonPlayerCharacter 
 {
-	protected int challengeRating;
+	protected float challengeRating;
 	
 }
